refactor: move utility functions out of preview generator service
